adding valve launch_testing pytest; had to split test to re-run fixture_server
