+++ buoy_tests/package.xml
@@ -9,6 +9,8 @@
   <author>Louise Poubel</author>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>ament_cmake_python</buildtool_depend>
+
   <buildtool_depend>rosidl_default_generators</buildtool_depend>
 
   <test_depend>buoy_gazebo</test_depend>
